+++ Simulation/Aivika/Arrival.hs
@@ -1,7 +1,7 @@
 
 -- |
 -- Module     : Simulation.Aivika.Arrival
--- Copyright  : Copyright (c) 2009-2016, David Sorokin <david.sorokin@gmail.com>
+-- Copyright  : Copyright (c) 2009-2017, David Sorokin <david.sorokin@gmail.com>
 -- License    : BSD3
 -- Maintainer : David Sorokin <david.sorokin@gmail.com>
 -- Stability  : experimental
@@ -22,7 +22,8 @@ module Simulation.Aivika.Arrival
         arrivalTimerChannel,
         arrivalProcessingTime,
         arrivalProcessingTimeChanged,
-        arrivalProcessingTimeChanged_) where
+        arrivalProcessingTimeChanged_,
+        resetArrivalTimer) where
 
 import Control.Monad
 import Control.Monad.Trans
@@ -102,3 +103,9 @@ arrivalTimerChannel :: ArrivalTimer -> Channel (Arrival a) (Arrival a)
 arrivalTimerChannel timer =
   Channel $ \sa ->
   return $ arrivalTimerSignal timer sa
+
+-- | Reset the statistics.
+resetArrivalTimer :: ArrivalTimer -> Event ()
+resetArrivalTimer timer =
+  do writeRef (arrivalProcessingTimeRef timer) emptySamplingStats
+     triggerSignal (arrivalProcessingTimeChangedSource timer) ()