+++ Simulation/Aivika/Parameter.hs
@@ -1,6 +1,6 @@
 -- |
 -- Module     : Simulation.Aivika.Parameter
--- Copyright  : Copyright (c) 2009-2016, David Sorokin <david.sorokin@gmail.com>
+-- Copyright  : Copyright (c) 2009-2017, David Sorokin <david.sorokin@gmail.com>
 -- License    : BSD3
 -- Maintainer : David Sorokin <david.sorokin@gmail.com>
 -- Stability  : experimental