+++ Simulation/Aivika/PriorityQueue/Pure.hs
@@ -1,7 +1,7 @@
 
 -- |
 -- Module     : Simulation.Aivika.PriorityQueue.Pure
--- Copyright  : Copyright (c) 2015-2016, David Sorokin <david.sorokin@gmail.com>
+-- Copyright  : Copyright (c) 2015-2017, David Sorokin <david.sorokin@gmail.com>
 -- License    : BSD3
 -- Maintainer : David Sorokin <david.sorokin@gmail.com>
 -- Stability  : experimental