+++ tests/Random/MainUsingCairo.hs
@@ -2,10 +2,11 @@
 -- To run, package aivika-experiment-cairo must be installed.
 
 import Simulation.Aivika.Experiment
+import Simulation.Aivika.Experiment.Chart
 import Simulation.Aivika.Experiment.Chart.Backend.Cairo
 
 import Graphics.Rendering.Chart.Backend.Cairo
 
 import Model
 
-main = runExperiment experiment generators (WebPageRenderer $ CairoRenderer PNG) model
+main = runExperiment experiment generators (WebPageRenderer (CairoRenderer PNG) experimentFilePath) model