+++ tests/Random/MainUsingDiagrams.hs
@@ -2,6 +2,7 @@
 -- To run, package aivika-experiment-diagrams must be installed.
 
 import Simulation.Aivika.Experiment
+import Simulation.Aivika.Experiment.Chart
 import Simulation.Aivika.Experiment.Chart.Backend.Diagrams
 
 import Graphics.Rendering.Chart.Backend.Diagrams
@@ -10,4 +11,7 @@ import qualified Data.Map as M
 
 import Model
 
-main = runExperiment experiment generators (WebPageRenderer $ DiagramsRenderer SVG loadCommonFonts) model
+main = 
+  do fonts <- loadCommonFonts
+     let renderer = DiagramsRenderer SVG (return fonts)
+     runExperiment experiment generators (WebPageRenderer renderer experimentFilePath) model